Rename ML modules and clean up code
