--- backend/server/ml_models/all_xgb/__init__.py
@@ -1,1 +1,0 @@
-from .all_xgb import AllXGB, AllXGBData
