--- backend/server/ml_models/betting_lasso/__init__.py
@@ -1,1 +1,0 @@
-from .betting_lasso import BettingLasso, BettingLassoData
