+++ backend/server/ml_models/betting_model/__init__.py
@@ -1,0 +1,1 @@
+from .betting_model import BettingModel, BettingModelData
