--- backend/server/ml_models/match_xgb/__init__.py
@@ -1,1 +1,0 @@
-from .match_xgb import MatchXGB, MatchXGBData
