+++ backend/server/ml_models/player_model/__init__.py
@@ -1,0 +1,1 @@
+from .player_model import PlayerModel, PlayerModelData
