--- backend/server/ml_models/player_xgb/__init__.py
@@ -1,1 +1,0 @@
-from .player_xgb import PlayerXGB, PlayerXGBData
