+++ backend/server/tests/unit/ml_models/__init__.py
@@ -1,0 +1,6 @@
+import sys
+
+from project.settings.common import BASE_DIR
+
+if BASE_DIR not in sys.path:
+    sys.path.append(BASE_DIR)
