+++ backend/server/tests/unit/ml_models/test_all_model.py
@@ -5,10 +5,11 @@ import numpy as np
 from faker import Faker
 
 from project.settings.common import BASE_DIR
-from server.ml_models import PlayerXGB
-from server.ml_models.player_xgb import PlayerXGBData
+from server.ml_models import AllModel
+from server.ml_models.all_model import AllModelData
 
 FAKE = Faker()
+N_ROWS = 10
 
 get_afltables_stats_df = pd.read_csv(
     f"{BASE_DIR}/server/tests/fixtures/fitzroy_get_afltables_stats.csv"
@@ -20,7 +21,7 @@ get_afltables_stats_mock = Mock(return_value=get_afltables_stats_df)
 match_results_mock = Mock(return_value=match_results_df)
 
 
-class TestPlayerXGB(TestCase):
+class TestAllModel(TestCase):
     def setUp(self):
         data_frame = pd.DataFrame(
             {
@@ -33,7 +34,7 @@ class TestPlayerXGB(TestCase):
         )
         self.X = pd.get_dummies(data_frame.drop("oppo_score", axis=1)).astype(float)
         self.y = data_frame["oppo_score"]
-        self.model = PlayerXGB()
+        self.model = AllModel()
 
     def test_predict(self):
         self.model.fit(self.X, self.y)
@@ -42,11 +43,52 @@ class TestPlayerXGB(TestCase):
         self.assertIsInstance(predictions, pd.Series)
 
 
-class TestPlayerXGBData(TestCase):
+class TestAllModelData(TestCase):
     def setUp(self):
-        self.data = PlayerXGBData(
-            data_readers=[get_afltables_stats_mock, match_results_mock]
-        )
+        teams = [FAKE.company() for _ in range(N_ROWS)]
+        years = ([2014] * 2) + ([2015] * 6) + ([2016] * 2)
+        round_numbers = 15
+        scores = np.random.randint(50, 150, N_ROWS)
+        oppo_scores = np.random.randint(50, 150, N_ROWS)
+        index_cols = ["team", "year", "round_number"]
+
+        betting_data = pd.DataFrame(
+            {
+                "team": teams,
+                "year": years,
+                "score": scores,
+                "oppo_score": oppo_scores,
+                "round_number": round_numbers,
+                "win_odds": (np.random.rand(N_ROWS) * 2) + 1,
+                "oppo_win_odds": (np.random.rand(N_ROWS) * 2) + 1,
+            }
+        ).set_index(index_cols, drop=False)
+
+        player_data = pd.DataFrame(
+            {
+                "team": teams,
+                "year": years,
+                "score": scores,
+                "oppo_score": oppo_scores,
+                "round_number": round_numbers,
+                "kicks": np.random.randint(1, 20, N_ROWS),
+                "marks": np.random.randint(1, 20, N_ROWS),
+            }
+        ).set_index(index_cols, drop=False)
+
+        match_data = pd.DataFrame(
+            {
+                "team": teams,
+                "year": years,
+                "score": scores,
+                "oppo_score": oppo_scores,
+                "round_number": round_numbers,
+                "rolling_win_percentage": np.random.rand(N_ROWS),
+                "ladder_position": np.random.randint(1, 18, N_ROWS),
+            }
+        ).set_index(index_cols, drop=False)
+
+        self.data = AllModelData(data_readers=[betting_data, player_data, match_data])
 
     def test_train_data(self):
         X_train, y_train = self.data.train_data()
@@ -55,10 +97,6 @@ class TestPlayerXGBData(TestCase):
         self.assertIsInstance(y_train, pd.Series)
         self.assertNotIn("score", X_train.columns)
         self.assertNotIn("oppo_score", X_train.columns)
-        self.assertNotIn("goals", X_train.columns)
-        self.assertNotIn("oppo_goals", X_train.columns)
-        self.assertNotIn("behinds", X_train.columns)
-        self.assertNotIn("oppo_behinds", X_train.columns)
         # No columns should be composed of strings
         self.assertFalse(
             any([X_train[column].dtype == "O" for column in X_train.columns])
@@ -75,10 +113,6 @@ class TestPlayerXGBData(TestCase):
         self.assertIsInstance(y_test, pd.Series)
         self.assertNotIn("score", X_test.columns)
         self.assertNotIn("oppo_score", X_test.columns)
-        self.assertNotIn("goals", X_test.columns)
-        self.assertNotIn("oppo_goals", X_test.columns)
-        self.assertNotIn("behinds", X_test.columns)
-        self.assertNotIn("oppo_behinds", X_test.columns)
         # No columns should be composed of strings
         self.assertFalse(
             any([X_test[column].dtype == "O" for column in X_test.columns])
