+++ backend/server/types.py
@@ -1,4 +1,9 @@
-from typing import Callable
+"""Module for custom static data types"""
+
+# pylint: disable=C0103
+
+from typing import Callable, Tuple, Optional
 import pandas as pd
 
 FeatureFunctionType = Callable[[pd.DataFrame], pd.DataFrame]
+YearsType = Tuple[Optional[int], Optional[int]]
